pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "asadaligith", email = "asad2747982@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlet>=1.3.1",
    "chainlit>=2.6.8",
    "dotenv>=0.9.9",
    "openai-agents>=0.2.4",
]

[project.scripts]
tools = "tools:main"
main = "tools.main:start"
chain = "tools.chain:start"
inputguard = "tools.input_guard:start"
outputguard = "tools.output_guard:start"
multiagent = "tools.multiagent:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
